web/requirements.txt
Django
gunicorn
django-haystack
elasticsearch
psycopg2
django-grappelli
django-tables2
django-bootstrap3
redis
